+++ lib/src/model/session_proposal.dart
@@ -10,9 +10,14 @@ class SessionProposal {
   final String id;
   final AppMetadata proposer;
   final Map<String, ProposalNamespace> namespaces;
+  @JsonKey(includeIfNull: false)
+  final Map<String, ProposalNamespace>? optionalNamespaces;
 
   SessionProposal(
-      {required this.id, required this.proposer, required this.namespaces});
+      {required this.id,
+      required this.proposer,
+      required this.namespaces,
+      this.optionalNamespaces});
 
   factory SessionProposal.fromJson(Map<String, dynamic> json) =>
       _$SessionProposalFromJson(json);